+++ src/main/java/ku_rum/backend/global/config/security/SecurityConfig.java
@@ -2,6 +2,7 @@ package ku_rum.backend.global.config.security;
 
 import ku_rum.backend.domain.user.domain.repository.UserRepository;
 import ku_rum.backend.global.config.redis.RedisUtil;
+import ku_rum.backend.global.response.status.BaseExceptionResponseStatus;
 import ku_rum.backend.global.security.jwt.CustomUserDetails;
 import ku_rum.backend.global.security.jwt.JwtTokenAuthenticationFilter;
 import ku_rum.backend.global.security.jwt.JwtTokenProvider;
@@ -23,6 +24,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+import static ku_rum.backend.global.response.status.BaseExceptionResponseStatus.*;
+
 @EnableWebSecurity
 @Configuration
 @RequiredArgsConstructor
@@ -36,8 +39,7 @@ public class SecurityConfig {
         http.formLogin(AbstractHttpConfigurer::disable);
         http.httpBasic(AbstractHttpConfigurer::disable);
         http.authorizeHttpRequests((auth) -> auth
-                .requestMatchers("/api/v1/auth/login", "/api/v1/users", "/docs/**", "/api/v1/users/validations/email"
-                        , "/api/v1/users/weinlogin").permitAll()
+                .requestMatchers(AuthorizationList.LIST.getAuthorities()).permitAll()
                 .anyRequest().authenticated());
         http.addFilterBefore(new JwtTokenAuthenticationFilter(jwtTokenProvider, redisUtil), UsernamePasswordAuthenticationFilter.class);
         http.sessionManagement((session) -> session
@@ -47,12 +49,12 @@ public class SecurityConfig {
 
     @Bean
     public UserDetailsService userDetailsService(UserRepository userRepository) {
-        return email -> userRepository.findUserByEmail(email)
+        return loginId -> userRepository.findUserByLoginId(loginId)
                 .map(u -> CustomUserDetails.of(u.getId(),
                         u.getEmail(),
                         AuthorityUtils.createAuthorityList(u.getRoles().toArray(new String[0])),
                         u.getPassword()))
-                .orElseThrow(() -> new UsernameNotFoundException("해당 유저가 존재하지 않습니다!"));
+                .orElseThrow(() -> new UsernameNotFoundException(NO_SUCH_USER.getMessage()));
     }
 
     @Bean
