+++ src/main/java/ku_rum/backend/global/security/jwt/JwtTokenProvider.java
@@ -4,6 +4,7 @@ import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import jakarta.annotation.PostConstruct;
 import ku_rum.backend.global.config.redis.RedisUtil;
+import ku_rum.backend.global.security.jwt.dto.TokenResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
